🔨Refactor: url 수정
ProfileController: 메서드 명칭 변경

Resolves: #12
