+++ src/main/kotlin/com/teamsparta/buysell/domain/member/controller/MemberController.kt
@@ -1,14 +1,11 @@
 package com.teamsparta.buysell.domain.member.controller
 
 import com.teamsparta.buysell.domain.member.dto.request.LoginRequest
-import com.teamsparta.buysell.domain.member.dto.request.MemberProfileUpdateRequest
 import com.teamsparta.buysell.domain.member.dto.request.SignUpRequest
-import com.teamsparta.buysell.domain.member.dto.response.MemberResponse
 import com.teamsparta.buysell.domain.member.service.MemberService
 import com.teamsparta.buysell.domain.member.service.SocialService
 import com.teamsparta.buysell.infra.security.UserPrincipal
 import com.teamsparta.buysell.infra.social.jwt.JwtDto
-import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
 import org.springframework.http.HttpHeaders
@@ -47,26 +44,6 @@ class MemberController(
     }
 
 
-    @PutMapping
-    @Operation(summary = "프로필 수정", description = "프로필을 수정합니다.")
-    fun updateProfile(
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-        @RequestBody request: MemberProfileUpdateRequest
-    ): ResponseEntity<MemberResponse> {
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(memberService.updateMyProfile(userPrincipal, request))
-    }
-
-    @GetMapping
-    @Operation(summary = "프로필 조회", description = "프로필을 조회합니다.")
-    fun getProfile(
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-    ): ResponseEntity<MemberResponse> {
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(memberService.getMyProfile(userPrincipal))
-    }
 
 
     @PutMapping("/signout")
