+++ src/main/kotlin/com/teamsparta/buysell/domain/member/service/MemberService.kt
@@ -1,7 +1,6 @@
 package com.teamsparta.buysell.domain.member.service
 
 import com.teamsparta.buysell.domain.member.dto.request.LoginRequest
-import com.teamsparta.buysell.domain.member.dto.request.MemberProfileUpdateRequest
 import com.teamsparta.buysell.domain.member.dto.request.SignUpRequest
 import com.teamsparta.buysell.domain.member.dto.response.MemberResponse
 import com.teamsparta.buysell.infra.security.UserPrincipal
@@ -9,8 +8,6 @@ import com.teamsparta.buysell.infra.security.UserPrincipal
 interface MemberService {
     fun signUp(request: SignUpRequest): MemberResponse
     fun login(request: LoginRequest): String
-    fun getMyProfile(userPrincipal: UserPrincipal): MemberResponse?
-    fun updateMyProfile(userPrincipal: UserPrincipal, request: MemberProfileUpdateRequest): MemberResponse
     fun signOut(userPrincipal: UserPrincipal)
     //fun googleLogin(oAuth2User: OAuth2User): JwtDto
 }
