+++ src/main/kotlin/com/teamsparta/buysell/domain/member/service/MemberServiceImpl.kt
@@ -2,7 +2,6 @@ package com.teamsparta.buysell.domain.member.service
 
 import com.teamsparta.buysell.domain.exception.ModelNotFoundException
 import com.teamsparta.buysell.domain.member.dto.request.LoginRequest
-import com.teamsparta.buysell.domain.member.dto.request.MemberProfileUpdateRequest
 import com.teamsparta.buysell.domain.member.dto.request.SignUpRequest
 import com.teamsparta.buysell.domain.member.dto.response.MemberResponse
 import com.teamsparta.buysell.domain.member.model.Account
@@ -17,7 +16,6 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.authentication.BadCredentialsException
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 
 @Service
 class MemberServiceImpl(
@@ -53,23 +51,7 @@ class MemberServiceImpl(
         return token
     }
 
-    // 현재 로그인 한 멤버 아이디 기준 정보 조회
-    @Transactional
-    override fun getMyProfile(userPrincipal: UserPrincipal): MemberResponse? {
-        val member = memberRepository.findByIdOrNull(userPrincipal.id)
-            ?: throw ModelNotFoundException("Member", userPrincipal.id)
-        return member.toResponse()
-    }
 
-    // 로그인 한 멤버 아이디 기준 정보 수정
-    @Transactional
-    override fun updateMyProfile(userPrincipal: UserPrincipal, request: MemberProfileUpdateRequest): MemberResponse {
-        val member = memberRepository.findByIdOrNull(userPrincipal.id)
-            ?: throw ModelNotFoundException("Member", userPrincipal.id)
-        member.nickname = request.nickname
-        member.birthday = request.birthday
-       return memberRepository.save(member).toResponse()
-    }
 
     // 멤버가 쓴 글 전체 조회
