+++ src/main/kotlin/com/teamsparta/buysell/domain/member/service/ProfileServiceImpl.kt
@@ -1,6 +1,8 @@
 package com.teamsparta.buysell.domain.member.service
 
 import com.teamsparta.buysell.domain.exception.ModelNotFoundException
+import com.teamsparta.buysell.domain.member.dto.request.MemberProfileUpdateRequest
+import com.teamsparta.buysell.domain.member.dto.response.MemberResponse
 import com.teamsparta.buysell.domain.member.repository.MemberRepository
 import com.teamsparta.buysell.domain.post.dto.response.PostResponse
 import com.teamsparta.buysell.domain.post.model.toResponse
@@ -9,6 +11,7 @@ import com.teamsparta.buysell.domain.post.repository.PostRepository
 import com.teamsparta.buysell.infra.security.UserPrincipal
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
 class ProfileServiceImpl(
@@ -30,6 +33,23 @@ class ProfileServiceImpl(
         val like = likeRepository.findByMember(member)
         val post = like.map { it.post }
         return post.map { it.toResponse() }
+    }
+    // 현재 로그인 한 멤버 아이디 기준 정보 조회
+    @Transactional
+    override fun getMyProfile(userPrincipal: UserPrincipal): MemberResponse? {
+        val member = memberRepository.findByIdOrNull(userPrincipal.id)
+            ?: throw ModelNotFoundException("Member", userPrincipal.id)
+        return member.toResponse()
+    }
+
+    // 로그인 한 멤버 아이디 기준 정보 수정
+    @Transactional
+    override fun updateMyProfile(userPrincipal: UserPrincipal, request: MemberProfileUpdateRequest): MemberResponse {
+        val member = memberRepository.findByIdOrNull(userPrincipal.id)
+            ?: throw ModelNotFoundException("Member", userPrincipal.id)
+        member.nickname = request.nickname
+        member.birthday = request.birthday
+        return memberRepository.save(member).toResponse()
     }
 
 }
